Update use of tax line items in sales receipts to use amount
- Update the call `adjustment_product_from_qb` to pass a third argument
for the tax_item. Without it, the method will use the tax name given in
the param instead of the line item
- Pass in the amount when creating the XML for the line when using taxes
- Create a method `sales_receipt_line_add_optional_rate` so that the
line will not add a rate if it doesn't see a price. This fixes the error
```
"You cannot set the the rate for the given sales tax item line, use the amount field instead"
```
I opted to create a new method instead of updating
`sales_receipt_line_add` because rate might be a required field in
another type. The current behaviour gives a rate of 0.00 even if a price
was never given.