Add a mapping for object_type to data object found and returned
The `CUSTOM_OBJECT_TYPES` lists the object_type it will look for and
`OBJECT_TYPES_MAPPING_DATA_OBJECT` list the data object it should
return. The metaprogramming expects the endpoint name to match the
endpoint name; `get_products` would give `products`. Since we're using
get endpoints that are unusual like `get_otherchargeproducts`, it'll try
to return `otherchargeproducts` which would not match a data object type
in Flowlink and would then be ignored.